+++ src/operator/tensor/broadcast_reduce_op_value.cc
@@ -27,192 +27,10 @@
 namespace mxnet {
 namespace op {
 DMLC_REGISTER_PARAMETER(ReduceAxesParam);
-DMLC_REGISTER_PARAMETER(NormParam);
 DMLC_REGISTER_PARAMETER(ReduceAxisParam);
 DMLC_REGISTER_PARAMETER(BroadcastAxesParam);
 DMLC_REGISTER_PARAMETER(BroadcastToParam);
 DMLC_REGISTER_PARAMETER(BroadcastLikeParam);
-
-inline std::string get_reduce_axes_description(const std::string& op_name, int line) {
-  std::string doc = R"code(Computes the __op__ of array elements over given axes.
-
-Defined in )code";
-  doc += std::string(__FILE__) + std::string(":L") + std::to_string(line);
-  size_t pos = 0;
-  std::string holder("__op__");
-  while ((pos = doc.find(holder, pos)) != std::string::npos) {
-    doc.replace(pos, holder.length(), op_name);
-    pos += op_name.length();
-  }
-  return doc;
-}
-
-template<>
-void L2NormComputeEx<cpu>(const nnvm::NodeAttrs& attrs,
-                          const OpContext& ctx,
-                          const std::vector<NDArray>& inputs,
-                          const std::vector<OpReqType>& req,
-                          const std::vector<NDArray>& outputs) {
-  CHECK_EQ(inputs.size(), 1U);
-  CHECK_EQ(outputs.size(), 1U);
-  CHECK_EQ(req.size(), 1U);
-  const NormParam& param = nnvm::get<NormParam>(attrs.parsed);
-  mshadow::Stream<cpu>* s = ctx.get_stream<cpu>();
-  const NDArrayStorageType istype = inputs[0].storage_type();
-  const mxnet::TShape axis = param.axis.has_value() ? param.axis.value() : mxnet::TShape();
-  if ((istype == kRowSparseStorage || istype == kCSRStorage) && axis.ndim() == 0 &&
-       param.ord == 2) {
-    // l2 norm on the entire array
-    L2NormComputeSparseImpl<cpu>(s, inputs[0], req[0], outputs[0].data());
-  } else if (istype == kCSRStorage && axis.ndim() == 1 && (axis[0] == 0 || axis[0] == 1) &&
-             !param.keepdims && param.ord == 2) {
-    // l2 norm on a particular axis
-    NDArray output = outputs[0];
-    ReduceCsrImpl<cpu, sq_sum, false>(s, ctx, inputs[0], req[0], &output, axis);
-    CHECK_EQ(outputs[0].storage_type(), kDefaultStorage);
-    SqRootForL2<cpu>(ctx, req[0], outputs[0].data());
-  } else {
-    LogUnimplementedOp(attrs, ctx, inputs, req, outputs);
-  }
-}
-
-MXNET_OPERATOR_REGISTER_REDUCE(sum)
-MXNET_ADD_SPARSE_OP_ALIAS(sum)
-.add_alias("sum_axis")
-.describe(R"code(Computes the sum of array elements over given axes.
-
-.. Note::
-
-  `sum` and `sum_axis` are equivalent.
-  For ndarray of csr storage type summation along axis 0 and axis 1 is supported.
-  Setting keepdims or exclude to True will cause a fallback to dense operator.
-
-Example::
-
-  data = [[[1, 2], [2, 3], [1, 3]],
-          [[1, 4], [4, 3], [5, 2]],
-          [[7, 1], [7, 2], [7, 3]]]
-
-  sum(data, axis=1)
-  [[  4.   8.]
-   [ 10.   9.]
-   [ 21.   6.]]
-
-  sum(data, axis=[1,2])
-  [ 12.  19.  27.]
-
-  data = [[1, 2, 0],
-          [3, 0, 1],
-          [4, 1, 0]]
-
-  csr = cast_storage(data, 'csr')
-
-  sum(csr, axis=0)
-  [ 8.  3.  1.]
-
-  sum(csr, axis=1)
-  [ 3.  4.  5.]
-
-)code" ADD_FILELINE)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow::red::sum>)
-.set_attr<FComputeEx>("FComputeEx<cpu>", ReduceAxesOpForwardEx<cpu, mshadow::red::sum>)
-.set_attr<FInferStorageType>("FInferStorageType", ReduceAxesOpForwardStorage)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ElemwiseGradUseNone{"_backward_sum"});
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_sum)
-.set_num_inputs(1)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseNone<cpu>);
-
-MXNET_OPERATOR_REGISTER_REDUCE(mean)
-MXNET_ADD_SPARSE_OP_ALIAS(mean)
-.describe(get_reduce_axes_description("mean", __LINE__))
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow::red::sum, true>)
-.set_attr<FComputeEx>("FComputeEx<cpu>", ReduceAxesOpForwardEx<cpu, mshadow::red::sum, true>)
-.set_attr<FInferStorageType>("FInferStorageType", ReduceAxesOpForwardStorage)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ElemwiseGradUseNone{"_backward_mean"});
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_mean)
-.set_num_inputs(1)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseNone<cpu, true>);
-
-MXNET_OPERATOR_REGISTER_REDUCE(prod)
-.describe(get_reduce_axes_description("product", __LINE__))
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow_op::product>)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ReduceGrad{ "_backward_prod" });
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_prod)
-.set_num_inputs(3)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseInOut< cpu, mshadow_op::rdiv>);
-
-MXNET_OPERATOR_REGISTER_REDUCE(nansum)
-.describe(R"code(Computes the sum of array elements over given axes treating Not a Numbers (``NaN``) as zero.
-
-)code" ADD_FILELINE)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow_op::nansum>)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ReduceGrad{ "_backward_nansum" });
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_nansum)
-.set_num_inputs(3)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseInOut<cpu, mshadow_op::nansum_grad>);
-
-MXNET_OPERATOR_REGISTER_REDUCE(nanprod)
-.describe(R"code(Computes the product of array elements over given axes treating Not a Numbers (``NaN``) as one.
-
-)code" ADD_FILELINE)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow_op::nanprod>)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ReduceGrad{ "_backward_nanprod" });
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_nanprod)
-.set_num_inputs(3)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseInOut<cpu, mshadow_op::nanprod_grad>);
-
-MXNET_OPERATOR_REGISTER_MINMAX_REDUCE(max)
-.add_alias("max_axis")
-.describe(get_reduce_axes_description("max", __LINE__))
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow::red::maximum>)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ReduceGrad{"_backward_max"});
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_max)
-.set_num_inputs(3)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseInOut<cpu, mshadow_op::eq>);
-
-MXNET_OPERATOR_REGISTER_MINMAX_REDUCE(min)
-.add_alias("min_axis")
-.describe(get_reduce_axes_description("min", __LINE__))
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesCompute<cpu, mshadow::red::minimum>)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<nnvm::FGradient>("FGradient", ReduceGrad{"_backward_min"});
-
-MXNET_OPERATOR_REGISTER_REDUCE_BACKWARD(_backward_min)
-.set_num_inputs(3)
-.set_attr<FCompute>("FCompute<cpu>", ReduceAxesBackwardUseInOut<cpu, mshadow_op::eq>);
 
 MXNET_OPERATOR_REGISTER_BROADCAST(broadcast_axis)
 .add_alias("broadcast_axes")
@@ -319,63 +137,6 @@ For example::
 .add_arguments(BroadcastLikeParam::__FIELDS__())
 .set_attr<mxnet::FInferShape>("FInferShape", BroadcastLikeShape)
 .set_attr<FCompute>("FCompute<cpu>", BroadcastCompute<cpu>);
-
-NNVM_REGISTER_OP(norm)
-MXNET_ADD_SPARSE_OP_ALIAS(norm)
-.describe(R"code(Computes the norm on an NDArray.
-
-This operator computes the norm on an NDArray with the specified axis, depending
-on the value of the ord parameter. By default, it computes the L2 norm on the entire
-array. Currently only ord=2 supports sparse ndarrays.
-
-Examples::
-
-  x = [[[1, 2],
-        [3, 4]],
-       [[2, 2],
-        [5, 6]]]
-
-  norm(x, ord=2, axis=1) = [[3.1622777 4.472136 ]
-                            [5.3851647 6.3245554]]
-
-  norm(x, ord=1, axis=1) = [[4., 6.],
-                            [7., 8.]]
-
-  rsp = x.cast_storage('row_sparse')
-
-  norm(rsp) = [5.47722578]
-
-  csr = x.cast_storage('csr')
-
-  norm(csr) = [5.47722578]
-
-)code" ADD_FILELINE)
-.set_num_inputs(1)
-.set_num_outputs(1)
-.set_attr_parser(ParamParser<NormParam>)
-.set_attr<mxnet::FInferShape>("FInferShape", NormShape)
-.set_attr<nnvm::FInferType>("FInferType", NormType)
-.set_attr<FInferStorageType>("FInferStorageType", LpNormStorageType)
-.set_attr<nnvm::FGradient>("FGradient", ReduceGrad{ "_backward_norm" })
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<FCompute>("FCompute<cpu>", LpNormCompute<cpu>)
-.set_attr<FComputeEx>("FComputeEx<cpu>", L2NormComputeEx<cpu>)
-.add_argument("data", "NDArray-or-Symbol", "The input")
-.add_arguments(NormParam::__FIELDS__());
-
-NNVM_REGISTER_OP(_backward_norm)
-.set_num_outputs(1)
-.set_attr_parser(ParamParser<NormParam>)
-.set_attr<nnvm::TIsBackward>("TIsBackward", true)
-.set_attr<FResourceRequest>("FResourceRequest",
-  [](const NodeAttrs& attrs) {
-    return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-  })
-.set_attr<FCompute>("FCompute<cpu>", LpNormGradCompute<cpu>);
-
 
 }  // namespace op
 }  // namespace mxnet
